Fix: Course adminCnt is too small

+++ src/core/models/Course.ts
@@ -76,7 +76,7 @@ export function init(courseDB: CourseDB) {
         allowNull: false,
       },
       admitCnt: {
-        type: DataTypes.TINYINT,
+        type: DataTypes.INTEGER,
         allowNull: false,
       },
       waitCnt: {
